requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
asyncpg
alembic
python-multipart
pytz
celery
redis
python-dotenv
